Mutulize game model call in sub module and start reworking   public ExtendedAgentStatCalculateModel(AgentStatCalculateModel previousModel)
