+++ Alliance.Editor/SubModule.cs
@@ -75,16 +75,6 @@ namespace Alliance.Editor
 			mission.AddMissionBehavior(new AdvancedCombatBehavior());
 		}
 
-		protected override void OnGameStart(Game game, IGameStarter gameStarter)
-		{
-			// Late patching, patching earlier causes issues with Voice type
-			Patch_AdvancedCombat.LatePatch();
-
-			// Add our custom GameModels 
-			gameStarter.AddModel(new ExtendedAgentStatCalculateModel());
-			gameStarter.AddModel(new ExtendedAgentApplyDamageModel());
-		}
-
 		private void GenerateScenarioExamples()
 		{
 			// Generate example scenario XML files
